requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.7
langchain-text-splitters==0.2.2

chromadb==0.5.5
pypdf==5.1.0
google-generativeai==0.7.2